# storage/api-docs/api-docs-v2.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "더 링코 코어(The Linko Core) Api 서비스",
        "description": "본 API 서비스는 자체 시비스인 차비스 를 위해 개발되었습니다.",
        "contact": {
            "email": "sh.jang@sohocode.kr"
        },
        "license": {
            "name": "Private"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "https://public.the-linko-core.wd/api/v2",
            "description": "백엔드 로컬 개발환경 전용"
        },
        {
            "url": "https://tlk.sohocode.kr/api/v2",
            "description": "프론트엔드 테스트서버"
        }
    ],
    "paths": {
        "/common/listCountry": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "국가정보 목록을 리턴한다.",
                "operationId": "ece26b26421f9f1441c4d590439865a5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/op"
                    },
                    {
                        "$ref": "#/components/parameters/keyword"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/page_per_items"
                    },
                    {
                        "$ref": "#/components/parameters/order"
                    },
                    {
                        "$ref": "#/components/parameters/dir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/page_info"
                                        },
                                        {
                                            "$ref": "#/components/schemas/country_list"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/getCountry/{id}": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "요청한 국가정보를 리턴한다.",
                "operationId": "e46248efd8f39de49e2e21a2d09e48af",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/country"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/listOccupationalGroup": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "직업군정보 목록을 리턴한다.",
                "operationId": "6f9fe077a9c367bfb8c89a0a6e5cc5df",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/op"
                    },
                    {
                        "$ref": "#/components/parameters/keyword"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/page_per_items"
                    },
                    {
                        "$ref": "#/components/parameters/order"
                    },
                    {
                        "$ref": "#/components/parameters/dir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/page_info"
                                        },
                                        {
                                            "$ref": "#/components/schemas/occupational_group_list"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/getOccupationalGroup/{id}": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "지정 직업군 정보를 리턴한다.",
                "operationId": "7fbe9db85099e704e0dc3142d539f7ef",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/occupational_group"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/listVisaDocumentType": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "비자발급시 필요한 문서 유형 목록을 출력한다.",
                "operationId": "6761fa2643559e635a9dc7f9494ce4af",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/visa_id"
                    },
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/op"
                    },
                    {
                        "$ref": "#/components/parameters/keyword"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/page_per_items"
                    },
                    {
                        "$ref": "#/components/parameters/order"
                    },
                    {
                        "$ref": "#/components/parameters/dir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/page_info"
                                        },
                                        {
                                            "$ref": "#/components/schemas/visa_document_types"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/getVisaDocumentType/{id}": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "비자발급시 필요한 지정 문서유형 정보를 출력한다.",
                "operationId": "3180f5a2644700a71828f3a3b5aa77cf",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/visa_document_type"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/listEvalInfo": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "지정 가능한 평가 설문 목록을 출력한다.",
                "operationId": "9252302fdb4529e2f780b099e34cebb0",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/target"
                    },
                    {
                        "$ref": "#/components/parameters/active"
                    },
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/op"
                    },
                    {
                        "$ref": "#/components/parameters/keyword"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/page_per_items"
                    },
                    {
                        "$ref": "#/components/parameters/order"
                    },
                    {
                        "$ref": "#/components/parameters/dir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/page_info"
                                        },
                                        {
                                            "$ref": "#/components/schemas/eval_infos"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/common/getEvaluation/{id}": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "지정 설문정보 일체를 출력한다.",
                "operationId": "cff5e58c525026b8042ad3824f658338",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요청 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/api_message"
                                        },
                                        {
                                            "$ref": "#/components/schemas/eval_info_include_item"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "api_message": {
                "title": "Api 메시지",
                "properties": {
                    "message": {
                        "description": "에러 메시지",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "page_info": {
                "title": "페이지정보",
                "properties": {
                    "total_items": {
                        "description": "총 검색 자료 수",
                        "type": "integer"
                    },
                    "total_page": {
                        "description": "총 검색 페이지 수",
                        "type": "integer"
                    },
                    "page": {
                        "description": "페이지 번호",
                        "type": "integer"
                    },
                    "page_per_items": {
                        "description": "페이지당 항목 수",
                        "type": "integer"
                    },
                    "item_count": {
                        "description": "실재 검색된 항목 수",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "country_list": {
                "title": "국가정보 목록",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/page_info"
                    },
                    {
                        "properties": {
                            "items": {
                                "description": "국가정보 목록",
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/country"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                ]
            },
...
            "occupational_group_list": {
                "title": "국가정보 목록",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/page_info"
                    },
                    {
                        "properties": {
                            "items": {
                                "description": "직업군정보 목록",
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/occupational_group"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                ]
            },
...
            "country": {
                "title": "국가정보",
                "properties": {
                    "id": {
                        "description": "일련번호",
                        "type": "integer"
                    },
                    "name": {
                        "description": "국가명",
                        "type": "string"
                    },
                    "en_name": {
                        "description": "국가명 (영문)",
                        "type": "string"
                    },
                    "code": {
                        "description": "국가코드",
                        "type": "string"
                    },
                    "iso3_code": {
                        "description": "ISO 3문자 국가코드",
                        "type": "string"
                    },
                    "nationality": {
                        "description": "국적",
                        "type": "string"
                    },
                    "continent": {
                        "description": "대륙명",
                        "type": "string"
                    },
                    "en_continent": {
                        "description": "대륙명 (영문)",
                        "type": "string"
                    },
                    "language_code": {
                        "description": "언어코드",
                        "type": "string"
                    },
                    "active": {
                        "description": "사용여부 (1:사용, 0:미사용)",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "404": {
                "description": "자료 없음 (찾지 못함)",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/api_message"
                        }
                    }
                }
            },
...
            "500": {
                "description": "서버 오류",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/api_message"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "id": {
                "name": "id",
                "in": "path",
                "description": "일련번호",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "visa_id": {
                "name": "visa_id",
                "in": "query",
                "description": "참고 대상 비자 신청 정보 일련번호",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
...
            "filter": {
                "name": "filter",
                "in": "query",
                "description": "필터 대상 필드명",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "op": {
                "name": "op",
                "in": "query",
                "description": "필터 연산자, 기본값 : like",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "like",
                        "=",
                        ">=",
                        "<=",
                        ">",
                        "<",
                        "<>"
                    ]
                }
            },
            "keyword": {
                "name": "keyword",
                "in": "query",
                "description": "필터 키워드",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "요청 페이지 번호, 기본값 : 1",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "page_per_items": {
                "name": "page_per_items",
                "in": "query",
                "description": "요청 페이지당 최대 항목 수, 기본값 : 50",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "order": {
                "name": "order",
                "in": "query",
                "description": "정렬 요청 필드",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "dir": {
                "name": "dir",
                "in": "query",
                "description": "정렬 방식, (asc, desc), 기본값 : asc",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            },
            "AccessTokenAuth": {
                "type": "apiKey",
                "name": "X-ACCESS-TOKEN",
                "in": "header"
            }
        }
    },
...
}